backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
PyJWT==2.10.1
python-multipart==0.0.20
pytest==8.3.4
